moved functions into global.ts
